tilemap: refactored TileMap class
Created two new classes: TileLoading which is responsible for loading
the tiles, and MapLoading which loads the Map. The only thing left in
TileMap class is setting the position of ???, gettig the size of the
tiles, and actualy drawing the tilemap.

tilemaps: refactored TileLoader and MapLoader

Refactored TileLoader, MapLoader, and TileMap according to comments from
previous push. Changed the names of the two classes from TileLoading and
MapLoading to TileLoader and MapLoader, changed order of varible
declarations so that privates are always at bottom, and fixed an issue
with the initialization of the MapLoader and TileLoader in TileMap
class.