entities/enemy: Character now takes damage
Added several functions to Berserker class that are responsible for
detecting if player has collided with enemy hitbox. Also, added two new
functions to CollisionDetection class: getRectangle and intersects that
create the rectangle hitbox around entities and check whether the
rectangle of the player intersected with the rectangle of the enemy.
When the character gets hit five times and his health reaches zero (in
the hit function of Berserker class) he gets send to the Main Menu of
the game.

Also added two functions that deal with player doing damage to enemies
but that still needs more work.