states/StateManager: refactored StateManager class
Created a new class StatePlaceholders that is responsible for
initializing the states ArrayList and is declared and used in the
StateManager class. Also refactored the setState function of the
StateManager class so that the initialization of the LevelOne class is
not happening in the constructor of the StateManager class but later.
This fixed a problem with player not being able to choose a character class.

states/StateManager: previous commit fixes

Fixed the issues from the previous pull request such as innecessary
Jpanel extension in CharState class, some confusing variable names, and
changed most variables in StatePlaceholder class to private.

entities/Player: player can now take damage/die

Added several functions to Player class: checkDamageTaken() which checks
if the player hitbox intersects with the enemy hitbox. Also added hit()
function which if there is a intersection then it substracts from the
player health and when that health hits 0 it takes us to the main menu
of the game.

Also added two functions to CollisionDetection class:
hitboxIntersection() which returns true if the hitbox of the player and enemy intersect, and
getEntityHitbox() which returns the dimenstions of the hitboxes around
entities.

Finaly, added a line to the update funtion in LevelOne class that checks
whether the player has taken any damage from the enemies on the map.