states/CharSelect: player can choose character
Modified the CharState class by adding a static string variable "choice" that tracks which character the player chose. This static variables is checked in LevelOne class and an appropriate hero is chosen to be loaded into the level. Also modified the
StateManager class by removing the state initializations in the
arraylist to a different function and not the constructor.