core/characters: multiple character selectio
Created a new class StatePlaceholders that is responsible for
initializing the states ArrayList and is declared and used in the
StateManager class. Also refactored the setState function of the
StateManager class so that the initialization of the LevelOne class is
not happening in the constructor of the StateManager class but later.
This fixed a problem with player not being able to choose a character class.

states/StateManager: previous commit fixes

Fixed the issues from the previous pull request such as innecessary
Jpanel extension in CharState class, some confusing variable names, and
changed most variables in StatePlaceholder class to private.