entities/Player: player can now take damage/die
Added several functions to Player class: checkDamageTaken() which checks
if the player hit box intersects with the enemy hit box. Also added
hit()
function which if there is a intersection then it subtracts from the
player health and when that health hits 0 it takes us to the main menu
of the game.

Also added two functions to CollisionDetection class:
hitboxIntersection() which returns true if the hit box of the player and
enemy intersect, and
getEntityHitbox() which returns the dimensions of the hit boxes around
entities.

Finally, added a line to the update function in LevelOne class that
checks
whether the player has taken any damage from the enemies on the map.